refactor(bundle-source): Divide implementations for formats
